// report-ng/app-react/src/model/report-model/framework_pb.ts
export type FailureCorridorValue = typeof FailureCorridorValue[keyof typeof FailureCorridorValue];

export namespace FailureCorridorValue {
  export type FCV_NOT_SET = typeof FailureCorridorValue.FCV_NOT_SET;
  export type FCV_HIGH = typeof FailureCorridorValue.FCV_HIGH;
  export type FCV_MID = typeof FailureCorridorValue.FCV_MID;
  export type FCV_LOW = typeof FailureCorridorValue.FCV_LOW;
}

export const MethodType = { MT_NOT_SET: 0, TEST_METHOD: 1, CONFIGURATION_METHOD: 2 } as const;

export type MethodType = typeof MethodType[keyof typeof MethodType];
// ...
export interface SuiteContext {
  contextValues?:
    | ContextValues
    | undefined;
  /**
   * list of all test
   *
   * @deprecated
   */
  testContextIds?:
    | string[]
    | undefined;
  /** reference */
  executionContextId?: string | undefined;
}
// ...
export interface TestContext {
  contextValues?:
    | ContextValues
    | undefined;
  /** @deprecated */
  classContextIds?: string[] | undefined;
  suiteContextId?:
    | string
    | undefined;
  /** @deprecated */
  executionContextId?: string | undefined;
}
// ...
export interface MethodContext {
  contextValues?: ContextValues | undefined;
  methodType?:
    | MethodType
    | undefined;
  /**
   * repeated string parameters = 8 [deprecated = true];
   *    repeated string method_tags = 9 [deprecated = true];
   */
  retryNumber?: number | undefined;
  methodRunIndex?: number | undefined;
  threadName?:
    | string
    | undefined;
  /** TestStep failed_step = 13; */
  failureCorridorValue?: FailureCorridorValue | undefined;
  classContextId?:
    | string
    | undefined;
  /** @deprecated */
  executionContextId?:
    | string
    | undefined;
  /**
   * repeated ErrorContext non_functional_infos = 17 [deprecated = true];
   *    repeated ErrorContext collected_assertions = 18 [deprecated = true];
   *
   * @deprecated
   */
  infos?:
    | string[]
    | undefined;
  /** @deprecated */
  priorityMessage?: string | undefined;
  relatedMethodContextIds?: string[] | undefined;
  dependsOnMethodContextIds?:
    | string[]
    | undefined;
  /** @deprecated */
  errorContext?: ErrorContext | undefined;
  testSteps?:
    | TestStep[]
    | undefined;
  /** @deprecated */
  testContextId?:
    | string
    | undefined;
  /** @deprecated */
  suiteContextId?: string | undefined;
  sessionContextIds?:
    | string[]
    | undefined;
  /**
   * repeated string video_ids = 30  [deprecated = true];
   *    repeated string screenshot_ids = 31 [deprecated = true];
   *    string custom_context_json = 32 [deprecated = true];
   */
  failedStepIndex?: number | undefined;
  resultStatus?: ResultStatusType | undefined;
  parameters?: { [key: string]: string } | undefined;
  customContexts?: { [key: string]: string } | undefined;
  annotations?:
    | { [key: string]: string }
    | undefined;
  /** A custom generated test name (e.a. cucumber scenario) */
  testName?: string | undefined;
  layoutCheckContext?: LayoutCheckContext[] | undefined;
}
// ...
export interface LogMessage {
  type?: LogMessageType | undefined;
  loggerName?: string | undefined;
  message?: string | undefined;
  timestamp?: number | undefined;
  threadName?: string | undefined;
  stackTrace?: StackTraceCause[] | undefined;
  prompt?: boolean | undefined;
  id?: string | undefined;
}

export interface ErrorContext {
  /**
   * string readable_error_message = 1 [deprecated = true];
   *    string additional_error_message = 2 [deprecated = true];
   *    StackTrace stack_trace = 3  [deprecated = true];
   *    string error_fingerprint = 6 [deprecated = true];
   */
  scriptSource?:
    | ScriptSource
    | undefined;
  /**
   * ScriptSource execution_object_source = 8 [deprecated = true];
   *
   * @deprecated
   */
  ticketId?:
    | string
    | undefined;
  /** @deprecated */
  description?: string | undefined;
  stackTrace?: StackTraceCause[] | undefined;
  optional?: boolean | undefined;
  id?: string | undefined;
}

export interface SessionContext {
  contextValues?:
    | ContextValues
    | undefined;
  /**
   * string session_key = 2 [deprecated = true];
   *    string provider = 3 [deprecated = true];
   *    map<string, string> metadata = 4 [deprecated = true];
   */
  sessionId?: string | undefined;
  videoId?: string | undefined;
  executionContextId?: string | undefined;
  browserName?: string | undefined;
  browserVersion?: string | undefined;
  capabilities?: string | undefined;
  serverUrl?: string | undefined;
  nodeUrl?: string | undefined;
  userAgent?: string | undefined;
  baseUrl?: string | undefined;
}
// ...
export const SuiteContext: MessageFns<SuiteContext> = {
  encode(message: SuiteContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.contextValues !== undefined) {
      ContextValues.encode(message.contextValues, writer.uint32(10).fork()).join();
    }
    if (message.testContextIds !== undefined && message.testContextIds.length !== 0) {
      for (const v of message.testContextIds) {
        writer.uint32(50).string(v!);
      }
    }
    if (message.executionContextId !== undefined && message.executionContextId !== "") {
      writer.uint32(58).string(message.executionContextId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SuiteContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSuiteContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.contextValues = ContextValues.decode(reader, reader.uint32());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          const el = reader.string();
          if (el !== undefined) {
            message.testContextIds!.push(el);
          }
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.executionContextId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const TestContext: MessageFns<TestContext> = {
  encode(message: TestContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.contextValues !== undefined) {
      ContextValues.encode(message.contextValues, writer.uint32(10).fork()).join();
    }
    if (message.classContextIds !== undefined && message.classContextIds.length !== 0) {
      for (const v of message.classContextIds) {
        writer.uint32(50).string(v!);
      }
    }
    if (message.suiteContextId !== undefined && message.suiteContextId !== "") {
      writer.uint32(58).string(message.suiteContextId);
    }
    if (message.executionContextId !== undefined && message.executionContextId !== "") {
      writer.uint32(66).string(message.executionContextId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TestContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTestContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.contextValues = ContextValues.decode(reader, reader.uint32());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          const el = reader.string();
          if (el !== undefined) {
            message.classContextIds!.push(el);
          }
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.suiteContextId = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.executionContextId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const SessionContext: MessageFns<SessionContext> = {
  encode(message: SessionContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.contextValues !== undefined) {
      ContextValues.encode(message.contextValues, writer.uint32(10).fork()).join();
    }
    if (message.sessionId !== undefined && message.sessionId !== "") {
      writer.uint32(50).string(message.sessionId);
    }
    if (message.videoId !== undefined && message.videoId !== "") {
      writer.uint32(58).string(message.videoId);
    }
    if (message.executionContextId !== undefined && message.executionContextId !== "") {
      writer.uint32(66).string(message.executionContextId);
    }
    if (message.browserName !== undefined && message.browserName !== "") {
      writer.uint32(74).string(message.browserName);
    }
    if (message.browserVersion !== undefined && message.browserVersion !== "") {
      writer.uint32(82).string(message.browserVersion);
    }
    if (message.capabilities !== undefined && message.capabilities !== "") {
      writer.uint32(90).string(message.capabilities);
    }
    if (message.serverUrl !== undefined && message.serverUrl !== "") {
      writer.uint32(98).string(message.serverUrl);
    }
    if (message.nodeUrl !== undefined && message.nodeUrl !== "") {
      writer.uint32(106).string(message.nodeUrl);
    }
    if (message.userAgent !== undefined && message.userAgent !== "") {
      writer.uint32(114).string(message.userAgent);
    }
    if (message.baseUrl !== undefined && message.baseUrl !== "") {
      writer.uint32(122).string(message.baseUrl);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SessionContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSessionContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.contextValues = ContextValues.decode(reader, reader.uint32());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.sessionId = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.videoId = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.executionContextId = reader.string();
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.browserName = reader.string();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.browserVersion = reader.string();
          continue;
        }
        case 11: {
          if (tag !== 90) {
            break;
          }

          message.capabilities = reader.string();
          continue;
        }
        case 12: {
          if (tag !== 98) {
            break;
          }

          message.serverUrl = reader.string();
          continue;
        }
        case 13: {
          if (tag !== 106) {
            break;
          }

          message.nodeUrl = reader.string();
          continue;
        }
        case 14: {
          if (tag !== 114) {
            break;
          }

          message.userAgent = reader.string();
          continue;
        }
        case 15: {
          if (tag !== 122) {
            break;
          }

          message.baseUrl = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};